Changing maximum size data to long to give more room for larger files

+++ api/src/main/java/org/openmrs/module/amrscustomization/AMRSCustomizationService.java
@@ -48,7 +48,7 @@ public interface AMRSCustomizationService {
     @Authorized({PrivilegeConstants.EDIT_PATIENTS})
     public List<MRNGeneratorLogEntry> getMRNGeneratorLogEntries();
 
-    public void setMaxUploadSize(Integer maxUploadSize);
+    public void setMaxUploadSize(Long maxUploadSize);
 
     public void reloadMaxUploadSize();
